Cleanup PluginList
