+++ js/app.js
@@ -22,7 +22,7 @@ import globalStyes from './global-styles';
 import App from './components/App';
 import TagList from './components/TagList.react';
 import NotFound from './components/NotFound';
-import PluginList from './components/PluginList.react';
+import PluginList from './components/PluginList';
 
 import rootReducer from './reducers/index';
 import Tags from './tags';
